refactor(repository): persistence restaurant 패키지 구조 변경
- Restaurant 관련 엔티티 및 리포지토리 import 경로를 com.reservation.persistence.restaurant.entity 로 이동/정리
- 여러 Restaurant*Entity 파일들의 package 선언을 com.reservation.persistence.restaurant.entity 로 변경
- mutator 및 jpa 리포지토리 파일들에서 변경된 패키지 구조에 맞춰 import 수정
- 일관된 패키지 네이밍으로 코드베이스 내 엔티티 참조 정합성 확보
