+++ adapter-module/src/main/kotlin/com/reservation/persistence/restaurant/entity/RestaurantCuisinesEntity.kt
@@ -1,4 +1,4 @@
-package com.reservation.persistence.restaurant
+package com.reservation.persistence.restaurant.entity
 
 import com.reservation.persistence.common.AuditDateTime
 import com.reservation.persistence.common.TimeBasedPrimaryKey
@@ -15,17 +15,17 @@ import org.hibernate.annotations.Comment
 
 @Table(
     catalog = "prototype_reservation",
-    name = "restaurant_tags",
+    name = "restaurant_cuisines",
     indexes = [
         Index(
-            columnList = "restaurant_id,tags_id",
+            columnList = "restaurant_id, cuisines_id",
             unique = false,
-            name = "index_restaurant_tags",
+            name = "index_restaurant_cuisines",
         ),
     ],
 )
 @Entity
-class RestaurantTagsEntity(
+class RestaurantCuisinesEntity(
     @ManyToOne(targetEntity = RestaurantEntity::class)
     @JoinColumn(
         name = "restaurant_id",
@@ -34,9 +34,9 @@ class RestaurantTagsEntity(
     )
     @Comment("음식점 식별자")
     private val restaurant: RestaurantEntity,
-    @Column(name = "tags_id")
+    @Column(name = "cuisines_id")
     @Comment("카테고리 식별자")
-    val tagsId: Long,
+    val cuisinesId: Long,
 ) : TimeBasedPrimaryKey() {
     @Embedded
     var auditDateTime: AuditDateTime = AuditDateTime()
