+++ adapter-module/src/main/kotlin/com/reservation/persistence/restaurant/entity/RestaurantWorkingDayId.kt
@@ -1,4 +1,4 @@
-package com.reservation.persistence.restaurant
+package com.reservation.persistence.restaurant.entity
 
 import jakarta.persistence.Column
 import jakarta.persistence.Embeddable
