+++ adapter-module/src/main/kotlin/com/reservation/persistence/restaurant/repository/jpa/RestaurantJpaRepository.kt
@@ -1,6 +1,6 @@
 package com.reservation.persistence.restaurant.repository.jpa
 
-import com.reservation.persistence.restaurant.RestaurantEntity
+import com.reservation.persistence.restaurant.entity.RestaurantEntity
 import org.springframework.data.repository.CrudRepository
 import java.util.Optional
